nbft: move spec types into separate header
To follow the existing pattern, move the NVMe specification types into a
separate header.
